combine breadcrumb component; add to App
