+++ frontend/src/pages/UI/Breadcrumb.jsx
@@ -1,6 +1,6 @@
 import { Link } from "react-router-dom";
 
-const Breadcrumb = (props) => {
+const BreadcrumbList = (props) => {
     return (
         <nav className="usa-breadcrumb" aria-label="Breadcrumbs,,">
             <ol className="usa-breadcrumb__list">
@@ -15,4 +15,18 @@ const Breadcrumb = (props) => {
     );
 };
 
-export default Breadcrumb;
+const BreadcrumbItem = (props) => {
+    return (
+        <li
+            className="usa-breadcrumb__list-item"
+            aria-label={props.pageName}
+            aria-current={props.isCurrent ? "page" : undefined}
+        >
+            <Link to="/portfolios" className="usa-breadcrumb__link">
+                {props.pageName}
+            </Link>
+        </li>
+    );
+};
+
+export { BreadcrumbList, BreadcrumbItem };
